+++ dotflow/__init__.py
@@ -2,15 +2,17 @@
 
 __version__ = "0.3.0"
 
-from .core.actions import Action as action
-from .core.actions import retry  # deprecated
-from .core.workflow import DotFlow
+from .core.action import Action as action
 from .core.context import Context
+from .core.workflow import DotFlow
+from .core.task import Task
+from .core.decorators import retry  # deprecated
 
 
 __all__ = [
     "action",
     "retry",
     "DotFlow",
-    "Context"
+    "Context",
+    "Task"
 ]