+++ tests/mocks.py
@@ -1,0 +1,32 @@
+"""Module Mock"""
+
+import logging
+
+from dotflow.core.action import Action as action
+
+
+@action
+def action_step() -> None:
+    pass
+
+
+@action
+def action_step_with_error() -> None:
+    raise Exception("Fail!")
+
+
+def simple_step() -> None:
+    pass
+
+
+def simple_callback(*args, **kwargs):
+    pass
+
+
+def dummy_step_with_fail() -> None:
+    logging.error("Fail!")
+    raise Exception("Fail!")
+
+
+def dummy_step_previous_context(previous_context):
+    logging.debug(previous_context.storage)